vana class 3: reverse-scan wrap loops + biblia sentinel
Backward loops on unsigned counters wrapped past zero instead of
stopping: selectio._est_ultimus_liberum (empty child list = wrap =
xar_obtinere NIHIL deref), stml retro/sandwich capture, biblia_visus
page scan, vindex stack walk (terminated only because inspicere
happened to fail on the wrapped index). Counters now s32 or
restructured to count-down-from-n idiom. biblia_invenire_librum gets
the tcp treatment: documented -1 now typed s32; dead clamps in
biblia_visus became live (page index on zero pages, dead < 0 on
lookup). stml_liberum_ad_indicem drops its dead index < 0 clause.

+++ lib/biblia_dr.c
@@ -73238,7 +73238,7 @@ _praefixum_aequale(constans character* textus, constans character* praefixum)
     redde VERUM;
 }
 
-i32
+s32
 biblia_invenire_librum(
     constans Biblia* biblia,
     constans character* quaesitum)
@@ -73251,7 +73251,7 @@ biblia_invenire_librum(
 
     si (!biblia || !quaesitum || !*quaesitum)
     {
-        redde (i32)(-I);
+        redde -I;
     }
 
     /* Videre si quaesitum incipit cum numero (e.g., "1 Maccabees") */
@@ -73295,7 +73295,7 @@ biblia_invenire_librum(
             {
                 si (_praefixum_aequale(nomen_anglicum + II, p))
                 {
-                    redde i;
+                    redde (s32)i;
                 }
             }
         }
@@ -73304,12 +73304,12 @@ biblia_invenire_librum(
             /* Comparatio simplex sine numero */
             si (_praefixum_aequale(nomen_anglicum, quaesitum))
             {
-                redde i;
+                redde (s32)i;
             }
         }
     }
 
-    redde (i32)(-I);
+    redde -I;
 }
 
 #if defined(__clang__)
